Perform reads with 6502 at 1 or 2MHz and pico at 200MHz
Increase time to wait before reading address_high by one pio
instruction cycle.

Use a hardcoded constant, tADDR, to time when to start capturing the address
after the start of the 6502 clock cycle.

Note: only tested with a R65C02. tADDR will need to be different for a
MOS 6502.

Could the pico detect the 6502 type by analysing the 6502 R_NW and
1MHZ signals at start up and tune/change the pio sm accordingly?